Cleanup hackathon unit tests (#55)
* Hessian --> hessian in unit test folders

* ConstraintHessian --> ConstraintsHessian in hiopkernels.

* UNIT_TEST_ --> UNIT_TESTS_.

* Capitalize test names for consistency.

* EQJAC --> EQUALITY_CONSTRAINT_JACOBIAN in test name.

* Include OPFLOW_ in related objective and constraints Jacobian unit test names.

* Cleanup a few CMake guards.

* Reorganize opflow unit tests.

* Remove  option from test_ineqjac_gpu, i.e, always evaluate performance.

* Minor edits to comments and print statements.

* Fix test_eqjac NETFILES and remove unnecessary command line option.

* Consolidate test_eqjac_gpu and test_eqjac_perf.

* Minor comment/print statement fix.

* Remove test_eqjac_perf from CMakeList.

---------
