+++ tests/unit/CMakeLists.txt
@@ -1,28 +1,6 @@
 add_subdirectory(opflow)
+add_subdirectory(pflow)
 add_subdirectory(ps)
-
-if(EXAGO_ENABLE_RAJA AND EXAGO_ENABLE_CUDA)
-  set_source_files_properties(test_acopf.cpp PROPERTIES LANGUAGE CUDA)
-endif()
-
-add_executable(test_acopf test_acopf.cpp utils/test_acopf_utils.cpp)
-target_link_libraries(test_acopf ExaGO::OPFLOW)
-target_include_directories(test_acopf PRIVATE ./utils)
-
-add_executable(test_pflow test_pflow.cpp utils/test_acopf_utils.cpp)
-target_link_libraries(test_pflow ExaGO::PFLOW)
-target_include_directories(test_pflow PRIVATE ./utils)
-
-if(EXAGO_ENABLE_RAJA AND EXAGO_ENABLE_HIOP_SPARSE)
-  if(EXAGO_ENABLE_CUDA)
-    set_source_files_properties(test_ineqjac_gpu.cpp PROPERTIES LANGUAGE CUDA)
-  endif()
-  add_executable(test_ineqjac_gpu test_ineqjac_gpu.cpp)
-  target_link_libraries(test_ineqjac_gpu ExaGO::OPFLOW)
-  target_include_directories(
-    test_ineqjac_gpu PRIVATE ${CMAKE_SOURCE_DIR}/src/opflow
-  )
-endif()
 
 add_executable(test_error_handler test_error_handler.cpp)
 target_link_libraries(test_error_handler ExaGO::UTILS)
@@ -31,51 +9,6 @@ add_executable(test_logger test_logger.cpp)
 target_link_libraries(test_logger ExaGO::UTILS)
 
 if(EXAGO_RUN_TESTS)
-  # Unit tests
-  exago_add_test(
-    NAME
-    "UNIT_TESTS_OPFLOW"
-    DEPENDS
-    HIOP
-    COMMAND
-    ${RUNCMD}
-    $<TARGET_FILE:test_acopf>
-    -opflow_genbusvoltage
-    VARIABLE_WITHIN_BOUNDS
-    -validation_dir
-    ${CMAKE_SOURCE_DIR}/datafiles/test_validation
-    NETFILES
-    ${network_files}
-  )
-
-  exago_add_test(
-    NAME
-    "UNIT_TESTS_PFLOW"
-    COMMAND
-    ${RUNCMD}
-    $<TARGET_FILE:test_pflow>
-    -validation_dir
-    ${CMAKE_SOURCE_DIR}/datafiles/test_validation
-    NETFILES
-    ${network_files}
-  )
-
-  if(EXAGO_ENABLE_RAJA AND EXAGO_ENABLE_HIOP_SPARSE)
-    exago_add_test(
-      NAME
-      "UNIT_TEST_INEQJAC_GPU"
-      DEPENDS
-      HIOP
-      COMMAND
-      ${RUNCMD}
-      $<TARGET_FILE:test_ineqjac_gpu>
-      -opflow_genbusvoltage
-      VARIABLE_WITHIN_BOUNDS
-      NETFILES
-      ${network_files}
-    )
-  endif()
-
   exago_add_test(
     NAME UNIT_TESTS_UTILS COMMAND $<TARGET_FILE:test_error_handler>
   )
@@ -92,5 +25,4 @@ if(EXAGO_RUN_TESTS)
     DEPENDS
     LOGGING
   )
-
 endif()
