+++ tests/unit/opflow/constraint_jacobian/inequality/CMakeLists.txt
@@ -65,7 +65,7 @@ if(EXAGO_RUN_TESTS)
     foreach(network num IN ZIP_LISTS obj_network_files num_copies)
       get_filename_component(net ${network} NAME)
       set(testname
-          "UNIT_TESTS_INEQUALITY_CONSTRAINT_JACOBIAN_${net}_${solver}_${model}"
+          "UNIT_TESTS_OPFLOW_INEQUALITY_CONSTRAINT_JACOBIAN_${net}_${solver}_${model}"
       )
       exago_add_test(
         NAME
