+++ tests/unit/opflow/hessian/README.md
@@ -14,4 +14,4 @@ A 5-bus system **CICJ-unittestx1.m** will be used as a basis for this test. In a
 ### Hessian sparsity structure:
 
 The Hessian has the following sparsity structure for our test:
-![Hessian_sparsity.png](/tests/unit/opflow/Hessian/Hessian_sparsity.png)
+![hessian_sparsity.png](/tests/unit/opflow/hessian/hessian_sparsity.png)
