+++ tests/unit/opflow/hessian/hessian_sparsity.png
